python/requirements.txt
# Kavach Titanium — Python Feeder Requirements
supabase==2.15.0
python-socketio[asyncio_client]==5.11.4
python-dotenv==1.0.1
numpy==1.26.4
pandas==2.2.3
scikit-learn==1.4.2
joblib==1.4.2
aiohttp==3.9.5
websockets==12.0
